utils/workflows/pdf-compare/requirements.txt
pymupdf==1.23.0
Pillow==12.2.0
